Pull `OpenOptions` out into its own file
We want to include it in documentation, and exports, but it can't go in
with `Upgrade` or `Database` or it will create a circular dependency.
