+++ docs/api/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"1":"External module","4":"Enumeration","16":"Enumeration member","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"helpers/PageComponentWrapper\"","url":"modules/_helpers_pagecomponentwrapper_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":256,"name":"PageComponent","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"helpers/PageComponentWrapper\""},{"id":2,"kind":1024,"name":"key","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":3,"kind":1024,"name":"Component","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#component","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":4,"kind":1024,"name":"props","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#props","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":5,"kind":256,"name":"PageComponentWrapper","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"helpers/PageComponentWrapper\""},{"id":6,"kind":1024,"name":"key","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponentWrapper"},{"id":7,"kind":2048,"name":"render","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html#render","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponentWrapper"},{"id":8,"kind":32,"name":"pageComponentWrapperPropType","url":"modules/_helpers_pagecomponentwrapper_.html#pagecomponentwrapperproptype","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"helpers/PageComponentWrapper\""},{"id":9,"kind":64,"name":"wrapPageComponent","url":"modules/_helpers_pagecomponentwrapper_.html#wrappagecomponent","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"helpers/PageComponentWrapper\""},{"id":10,"kind":1,"name":"\"components/Page\"","url":"modules/_components_page_.html","classes":"tsd-kind-external-module"},{"id":11,"kind":256,"name":"PageProps","url":"interfaces/_components_page_.pageprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Page\""},{"id":12,"kind":1024,"name":"componentWrappers","url":"interfaces/_components_page_.pageprops.html#componentwrappers","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Page\".PageProps"},{"id":13,"kind":128,"name":"Page","url":"classes/_components_page_.page.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"components/Page\""},{"id":14,"kind":2097152,"name":"propTypes","url":"classes/_components_page_.page.html#proptypes","classes":"tsd-kind-object-literal tsd-parent-kind-class tsd-is-static","parent":"\"components/Page\".Page"},{"id":15,"kind":32,"name":"componentWrappers","url":"classes/_components_page_.page.html#proptypes.componentwrappers","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Page\".Page.propTypes"},{"id":16,"kind":2048,"name":"render","url":"classes/_components_page_.page.html#render","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-overwrite","parent":"\"components/Page\".Page"},{"id":17,"kind":2048,"name":"componentDidMount","url":"classes/_components_page_.page.html#componentdidmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":18,"kind":2048,"name":"shouldComponentUpdate","url":"classes/_components_page_.page.html#shouldcomponentupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":19,"kind":2048,"name":"componentWillUnmount","url":"classes/_components_page_.page.html#componentwillunmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":20,"kind":2048,"name":"componentDidCatch","url":"classes/_components_page_.page.html#componentdidcatch","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":21,"kind":2048,"name":"getSnapshotBeforeUpdate","url":"classes/_components_page_.page.html#getsnapshotbeforeupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":22,"kind":2048,"name":"componentDidUpdate","url":"classes/_components_page_.page.html#componentdidupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":23,"kind":2048,"name":"componentWillMount","url":"classes/_components_page_.page.html#componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":24,"kind":2048,"name":"UNSAFE_componentWillMount","url":"classes/_components_page_.page.html#unsafe_componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":25,"kind":2048,"name":"componentWillReceiveProps","url":"classes/_components_page_.page.html#componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":26,"kind":2048,"name":"UNSAFE_componentWillReceiveProps","url":"classes/_components_page_.page.html#unsafe_componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":27,"kind":2048,"name":"componentWillUpdate","url":"classes/_components_page_.page.html#componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":28,"kind":2048,"name":"UNSAFE_componentWillUpdate","url":"classes/_components_page_.page.html#unsafe_componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":29,"kind":1024,"name":"contextType","url":"classes/_components_page_.page.html#contexttype","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-static","parent":"\"components/Page\".Page"},{"id":30,"kind":1024,"name":"context","url":"classes/_components_page_.page.html#context","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":31,"kind":512,"name":"constructor","url":"classes/_components_page_.page.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":32,"kind":2048,"name":"setState","url":"classes/_components_page_.page.html#setstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":33,"kind":2048,"name":"forceUpdate","url":"classes/_components_page_.page.html#forceupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":34,"kind":1024,"name":"props","url":"classes/_components_page_.page.html#props","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":35,"kind":1024,"name":"state","url":"classes/_components_page_.page.html#state","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":36,"kind":1024,"name":"refs","url":"classes/_components_page_.page.html#refs","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":37,"kind":65536,"name":"__type","url":"classes/_components_page_.page.html#refs.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"components/Page\".Page.refs"},{"id":38,"kind":1,"name":"\"components/Orchestrator\"","url":"modules/_components_orchestrator_.html","classes":"tsd-kind-external-module"},{"id":39,"kind":256,"name":"Step","url":"interfaces/_components_orchestrator_.step.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":40,"kind":1024,"name":"key","url":"interfaces/_components_orchestrator_.step.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".Step"},{"id":41,"kind":1024,"name":"componentWrappers","url":"interfaces/_components_orchestrator_.step.html#componentwrappers","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".Step"},{"id":42,"kind":256,"name":"OrchestratorProps","url":"interfaces/_components_orchestrator_.orchestratorprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":43,"kind":1024,"name":"currentStepKey","url":"interfaces/_components_orchestrator_.orchestratorprops.html#currentstepkey","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".OrchestratorProps"},{"id":44,"kind":1024,"name":"steps","url":"interfaces/_components_orchestrator_.orchestratorprops.html#steps","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".OrchestratorProps"},{"id":45,"kind":128,"name":"Orchestrator","url":"classes/_components_orchestrator_.orchestrator.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":46,"kind":2097152,"name":"propTypes","url":"classes/_components_orchestrator_.orchestrator.html#proptypes","classes":"tsd-kind-object-literal tsd-parent-kind-class tsd-is-static","parent":"\"components/Orchestrator\".Orchestrator"},{"id":47,"kind":32,"name":"currentStepKey","url":"classes/_components_orchestrator_.orchestrator.html#proptypes.currentstepkey","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Orchestrator\".Orchestrator.propTypes"},{"id":48,"kind":32,"name":"steps","url":"classes/_components_orchestrator_.orchestrator.html#proptypes.steps","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Orchestrator\".Orchestrator.propTypes"},{"id":49,"kind":2048,"name":"render","url":"classes/_components_orchestrator_.orchestrator.html#render","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-overwrite","parent":"\"components/Orchestrator\".Orchestrator"},{"id":50,"kind":2048,"name":"componentDidMount","url":"classes/_components_orchestrator_.orchestrator.html#componentdidmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":51,"kind":2048,"name":"shouldComponentUpdate","url":"classes/_components_orchestrator_.orchestrator.html#shouldcomponentupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":52,"kind":2048,"name":"componentWillUnmount","url":"classes/_components_orchestrator_.orchestrator.html#componentwillunmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":53,"kind":2048,"name":"componentDidCatch","url":"classes/_components_orchestrator_.orchestrator.html#componentdidcatch","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":54,"kind":2048,"name":"getSnapshotBeforeUpdate","url":"classes/_components_orchestrator_.orchestrator.html#getsnapshotbeforeupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":55,"kind":2048,"name":"componentDidUpdate","url":"classes/_components_orchestrator_.orchestrator.html#componentdidupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":56,"kind":2048,"name":"componentWillMount","url":"classes/_components_orchestrator_.orchestrator.html#componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":57,"kind":2048,"name":"UNSAFE_componentWillMount","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":58,"kind":2048,"name":"componentWillReceiveProps","url":"classes/_components_orchestrator_.orchestrator.html#componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":59,"kind":2048,"name":"UNSAFE_componentWillReceiveProps","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":60,"kind":2048,"name":"componentWillUpdate","url":"classes/_components_orchestrator_.orchestrator.html#componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":61,"kind":2048,"name":"UNSAFE_componentWillUpdate","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":62,"kind":1024,"name":"contextType","url":"classes/_components_orchestrator_.orchestrator.html#contexttype","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-static","parent":"\"components/Orchestrator\".Orchestrator"},{"id":63,"kind":1024,"name":"context","url":"classes/_components_orchestrator_.orchestrator.html#context","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":64,"kind":512,"name":"constructor","url":"classes/_components_orchestrator_.orchestrator.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":65,"kind":2048,"name":"setState","url":"classes/_components_orchestrator_.orchestrator.html#setstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":66,"kind":2048,"name":"forceUpdate","url":"classes/_components_orchestrator_.orchestrator.html#forceupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":67,"kind":1024,"name":"props","url":"classes/_components_orchestrator_.orchestrator.html#props","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":68,"kind":1024,"name":"state","url":"classes/_components_orchestrator_.orchestrator.html#state","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":69,"kind":1024,"name":"refs","url":"classes/_components_orchestrator_.orchestrator.html#refs","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":70,"kind":65536,"name":"__type","url":"classes/_components_orchestrator_.orchestrator.html#refs.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"components/Orchestrator\".Orchestrator.refs"},{"id":71,"kind":1,"name":"\"store/types\"","url":"modules/_store_types_.html","classes":"tsd-kind-external-module"},{"id":72,"kind":256,"name":"StoreSchema","url":"interfaces/_store_types_.storeschema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"store/types\""},{"id":73,"kind":1024,"name":"key","url":"interfaces/_store_types_.storeschema.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":74,"kind":1024,"name":"value","url":"interfaces/_store_types_.storeschema.html#value","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":75,"kind":1024,"name":"indexes","url":"interfaces/_store_types_.storeschema.html#indexes","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":76,"kind":256,"name":"Schema","url":"interfaces/_store_types_.schema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"store/types\""},{"id":77,"kind":256,"name":"NamedSchema","url":"interfaces/_store_types_.namedschema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":78,"kind":1024,"name":"dbNames","url":"interfaces/_store_types_.namedschema.html#dbnames","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".NamedSchema"},{"id":79,"kind":1024,"name":"schema","url":"interfaces/_store_types_.namedschema.html#schema","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".NamedSchema"},{"id":80,"kind":4194304,"name":"StoreNames","url":"modules/_store_types_.html#storenames","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":81,"kind":4194304,"name":"StoreKey","url":"modules/_store_types_.html#storekey","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":82,"kind":4194304,"name":"StoreValue","url":"modules/_store_types_.html#storevalue","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":83,"kind":4194304,"name":"Transaction","url":"modules/_store_types_.html#transaction","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":84,"kind":4194304,"name":"Store","url":"modules/_store_types_.html#store","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":85,"kind":4194304,"name":"StoreMap","url":"modules/_store_types_.html#storemap","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":86,"kind":65536,"name":"__type","url":"modules/_store_types_.html#storemap.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"\"store/types\".StoreMap"},{"id":87,"kind":1,"name":"\"store/Upgrade\"","url":"modules/_store_upgrade_.html","classes":"tsd-kind-external-module"},{"id":88,"kind":128,"name":"Upgrade","url":"classes/_store_upgrade_.upgrade.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/Upgrade\""},{"id":89,"kind":1024,"name":"oldVersion","url":"classes/_store_upgrade_.upgrade.html#oldversion","classes":"tsd-kind-property tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":90,"kind":1024,"name":"newVersion","url":"classes/_store_upgrade_.upgrade.html#newversion","classes":"tsd-kind-property tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":91,"kind":512,"name":"constructor","url":"classes/_store_upgrade_.upgrade.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":92,"kind":2048,"name":"createStore","url":"classes/_store_upgrade_.upgrade.html#createstore","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Upgrade\".Upgrade"},{"id":93,"kind":2048,"name":"transaction","url":"classes/_store_upgrade_.upgrade.html#transaction","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Upgrade\".Upgrade"},{"id":94,"kind":1,"name":"\"store/Database\"","url":"modules/_store_database_.html","classes":"tsd-kind-external-module"},{"id":95,"kind":4,"name":"TransactionMode","url":"enums/_store_database_.transactionmode.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"store/Database\""},{"id":96,"kind":16,"name":"ReadOnly","url":"enums/_store_database_.transactionmode.html#readonly","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"store/Database\".TransactionMode"},{"id":97,"kind":16,"name":"ReadWrite","url":"enums/_store_database_.transactionmode.html#readwrite","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"store/Database\".TransactionMode"},{"id":98,"kind":128,"name":"Database","url":"classes/_store_database_.database.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/Database\""},{"id":99,"kind":2048,"name":"open","url":"classes/_store_database_.database.html#open","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-static","parent":"\"store/Database\".Database"},{"id":100,"kind":2048,"name":"put","url":"classes/_store_database_.database.html#put","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":101,"kind":2048,"name":"get","url":"classes/_store_database_.database.html#get","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":102,"kind":2048,"name":"transaction","url":"classes/_store_database_.database.html#transaction","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":103,"kind":2048,"name":"close","url":"classes/_store_database_.database.html#close","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"store/Database\".Database"}]};
+            typedoc.search.data = {"kinds":{"1":"External module","4":"Enumeration","16":"Enumeration member","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"helpers/PageComponentWrapper\"","url":"modules/_helpers_pagecomponentwrapper_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":256,"name":"PageComponent","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"helpers/PageComponentWrapper\""},{"id":2,"kind":1024,"name":"key","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":3,"kind":1024,"name":"Component","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#component","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":4,"kind":1024,"name":"props","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponent.html#props","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponent"},{"id":5,"kind":256,"name":"PageComponentWrapper","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"helpers/PageComponentWrapper\""},{"id":6,"kind":1024,"name":"key","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponentWrapper"},{"id":7,"kind":2048,"name":"render","url":"interfaces/_helpers_pagecomponentwrapper_.pagecomponentwrapper.html#render","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"\"helpers/PageComponentWrapper\".PageComponentWrapper"},{"id":8,"kind":32,"name":"pageComponentWrapperPropType","url":"modules/_helpers_pagecomponentwrapper_.html#pagecomponentwrapperproptype","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"helpers/PageComponentWrapper\""},{"id":9,"kind":64,"name":"wrapPageComponent","url":"modules/_helpers_pagecomponentwrapper_.html#wrappagecomponent","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"helpers/PageComponentWrapper\""},{"id":10,"kind":1,"name":"\"components/Page\"","url":"modules/_components_page_.html","classes":"tsd-kind-external-module"},{"id":11,"kind":256,"name":"PageProps","url":"interfaces/_components_page_.pageprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Page\""},{"id":12,"kind":1024,"name":"componentWrappers","url":"interfaces/_components_page_.pageprops.html#componentwrappers","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Page\".PageProps"},{"id":13,"kind":128,"name":"Page","url":"classes/_components_page_.page.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"components/Page\""},{"id":14,"kind":2097152,"name":"propTypes","url":"classes/_components_page_.page.html#proptypes","classes":"tsd-kind-object-literal tsd-parent-kind-class tsd-is-static","parent":"\"components/Page\".Page"},{"id":15,"kind":32,"name":"componentWrappers","url":"classes/_components_page_.page.html#proptypes.componentwrappers","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Page\".Page.propTypes"},{"id":16,"kind":2048,"name":"render","url":"classes/_components_page_.page.html#render","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-overwrite","parent":"\"components/Page\".Page"},{"id":17,"kind":2048,"name":"componentDidMount","url":"classes/_components_page_.page.html#componentdidmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":18,"kind":2048,"name":"shouldComponentUpdate","url":"classes/_components_page_.page.html#shouldcomponentupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":19,"kind":2048,"name":"componentWillUnmount","url":"classes/_components_page_.page.html#componentwillunmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":20,"kind":2048,"name":"componentDidCatch","url":"classes/_components_page_.page.html#componentdidcatch","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":21,"kind":2048,"name":"getSnapshotBeforeUpdate","url":"classes/_components_page_.page.html#getsnapshotbeforeupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":22,"kind":2048,"name":"componentDidUpdate","url":"classes/_components_page_.page.html#componentdidupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":23,"kind":2048,"name":"componentWillMount","url":"classes/_components_page_.page.html#componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":24,"kind":2048,"name":"UNSAFE_componentWillMount","url":"classes/_components_page_.page.html#unsafe_componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":25,"kind":2048,"name":"componentWillReceiveProps","url":"classes/_components_page_.page.html#componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":26,"kind":2048,"name":"UNSAFE_componentWillReceiveProps","url":"classes/_components_page_.page.html#unsafe_componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":27,"kind":2048,"name":"componentWillUpdate","url":"classes/_components_page_.page.html#componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":28,"kind":2048,"name":"UNSAFE_componentWillUpdate","url":"classes/_components_page_.page.html#unsafe_componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":29,"kind":1024,"name":"contextType","url":"classes/_components_page_.page.html#contexttype","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-static","parent":"\"components/Page\".Page"},{"id":30,"kind":1024,"name":"context","url":"classes/_components_page_.page.html#context","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":31,"kind":512,"name":"constructor","url":"classes/_components_page_.page.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":32,"kind":2048,"name":"setState","url":"classes/_components_page_.page.html#setstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":33,"kind":2048,"name":"forceUpdate","url":"classes/_components_page_.page.html#forceupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":34,"kind":1024,"name":"props","url":"classes/_components_page_.page.html#props","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":35,"kind":1024,"name":"state","url":"classes/_components_page_.page.html#state","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":36,"kind":1024,"name":"refs","url":"classes/_components_page_.page.html#refs","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Page\".Page"},{"id":37,"kind":65536,"name":"__type","url":"classes/_components_page_.page.html#refs.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"components/Page\".Page.refs"},{"id":38,"kind":1,"name":"\"components/Orchestrator\"","url":"modules/_components_orchestrator_.html","classes":"tsd-kind-external-module"},{"id":39,"kind":256,"name":"Step","url":"interfaces/_components_orchestrator_.step.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":40,"kind":1024,"name":"key","url":"interfaces/_components_orchestrator_.step.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".Step"},{"id":41,"kind":1024,"name":"componentWrappers","url":"interfaces/_components_orchestrator_.step.html#componentwrappers","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".Step"},{"id":42,"kind":256,"name":"OrchestratorProps","url":"interfaces/_components_orchestrator_.orchestratorprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":43,"kind":1024,"name":"currentStepKey","url":"interfaces/_components_orchestrator_.orchestratorprops.html#currentstepkey","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".OrchestratorProps"},{"id":44,"kind":1024,"name":"steps","url":"interfaces/_components_orchestrator_.orchestratorprops.html#steps","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Orchestrator\".OrchestratorProps"},{"id":45,"kind":128,"name":"Orchestrator","url":"classes/_components_orchestrator_.orchestrator.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"components/Orchestrator\""},{"id":46,"kind":2097152,"name":"propTypes","url":"classes/_components_orchestrator_.orchestrator.html#proptypes","classes":"tsd-kind-object-literal tsd-parent-kind-class tsd-is-static","parent":"\"components/Orchestrator\".Orchestrator"},{"id":47,"kind":32,"name":"currentStepKey","url":"classes/_components_orchestrator_.orchestrator.html#proptypes.currentstepkey","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Orchestrator\".Orchestrator.propTypes"},{"id":48,"kind":32,"name":"steps","url":"classes/_components_orchestrator_.orchestrator.html#proptypes.steps","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"components/Orchestrator\".Orchestrator.propTypes"},{"id":49,"kind":2048,"name":"render","url":"classes/_components_orchestrator_.orchestrator.html#render","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-overwrite","parent":"\"components/Orchestrator\".Orchestrator"},{"id":50,"kind":2048,"name":"componentDidMount","url":"classes/_components_orchestrator_.orchestrator.html#componentdidmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":51,"kind":2048,"name":"shouldComponentUpdate","url":"classes/_components_orchestrator_.orchestrator.html#shouldcomponentupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":52,"kind":2048,"name":"componentWillUnmount","url":"classes/_components_orchestrator_.orchestrator.html#componentwillunmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":53,"kind":2048,"name":"componentDidCatch","url":"classes/_components_orchestrator_.orchestrator.html#componentdidcatch","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":54,"kind":2048,"name":"getSnapshotBeforeUpdate","url":"classes/_components_orchestrator_.orchestrator.html#getsnapshotbeforeupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":55,"kind":2048,"name":"componentDidUpdate","url":"classes/_components_orchestrator_.orchestrator.html#componentdidupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":56,"kind":2048,"name":"componentWillMount","url":"classes/_components_orchestrator_.orchestrator.html#componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":57,"kind":2048,"name":"UNSAFE_componentWillMount","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillmount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":58,"kind":2048,"name":"componentWillReceiveProps","url":"classes/_components_orchestrator_.orchestrator.html#componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":59,"kind":2048,"name":"UNSAFE_componentWillReceiveProps","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillreceiveprops","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":60,"kind":2048,"name":"componentWillUpdate","url":"classes/_components_orchestrator_.orchestrator.html#componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":61,"kind":2048,"name":"UNSAFE_componentWillUpdate","url":"classes/_components_orchestrator_.orchestrator.html#unsafe_componentwillupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":62,"kind":1024,"name":"contextType","url":"classes/_components_orchestrator_.orchestrator.html#contexttype","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-static","parent":"\"components/Orchestrator\".Orchestrator"},{"id":63,"kind":1024,"name":"context","url":"classes/_components_orchestrator_.orchestrator.html#context","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":64,"kind":512,"name":"constructor","url":"classes/_components_orchestrator_.orchestrator.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":65,"kind":2048,"name":"setState","url":"classes/_components_orchestrator_.orchestrator.html#setstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":66,"kind":2048,"name":"forceUpdate","url":"classes/_components_orchestrator_.orchestrator.html#forceupdate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":67,"kind":1024,"name":"props","url":"classes/_components_orchestrator_.orchestrator.html#props","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":68,"kind":1024,"name":"state","url":"classes/_components_orchestrator_.orchestrator.html#state","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":69,"kind":1024,"name":"refs","url":"classes/_components_orchestrator_.orchestrator.html#refs","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"\"components/Orchestrator\".Orchestrator"},{"id":70,"kind":65536,"name":"__type","url":"classes/_components_orchestrator_.orchestrator.html#refs.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"components/Orchestrator\".Orchestrator.refs"},{"id":71,"kind":1,"name":"\"store/types\"","url":"modules/_store_types_.html","classes":"tsd-kind-external-module"},{"id":72,"kind":256,"name":"StoreSchema","url":"interfaces/_store_types_.storeschema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"store/types\""},{"id":73,"kind":1024,"name":"key","url":"interfaces/_store_types_.storeschema.html#key","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":74,"kind":1024,"name":"value","url":"interfaces/_store_types_.storeschema.html#value","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":75,"kind":1024,"name":"indexes","url":"interfaces/_store_types_.storeschema.html#indexes","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".StoreSchema"},{"id":76,"kind":256,"name":"Schema","url":"interfaces/_store_types_.schema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"store/types\""},{"id":77,"kind":256,"name":"NamedSchema","url":"interfaces/_store_types_.namedschema.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":78,"kind":1024,"name":"dbNames","url":"interfaces/_store_types_.namedschema.html#dbnames","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".NamedSchema"},{"id":79,"kind":1024,"name":"schema","url":"interfaces/_store_types_.namedschema.html#schema","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/types\".NamedSchema"},{"id":80,"kind":4194304,"name":"StoreNames","url":"modules/_store_types_.html#storenames","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":81,"kind":4194304,"name":"StoreKey","url":"modules/_store_types_.html#storekey","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":82,"kind":4194304,"name":"StoreValue","url":"modules/_store_types_.html#storevalue","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":83,"kind":4194304,"name":"Transaction","url":"modules/_store_types_.html#transaction","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":84,"kind":4194304,"name":"Store","url":"modules/_store_types_.html#store","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":85,"kind":4194304,"name":"StoreMap","url":"modules/_store_types_.html#storemap","classes":"tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/types\""},{"id":86,"kind":65536,"name":"__type","url":"modules/_store_types_.html#storemap.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"\"store/types\".StoreMap"},{"id":87,"kind":1,"name":"\"store/Upgrade\"","url":"modules/_store_upgrade_.html","classes":"tsd-kind-external-module"},{"id":88,"kind":128,"name":"Upgrade","url":"classes/_store_upgrade_.upgrade.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/Upgrade\""},{"id":89,"kind":1024,"name":"oldVersion","url":"classes/_store_upgrade_.upgrade.html#oldversion","classes":"tsd-kind-property tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":90,"kind":1024,"name":"newVersion","url":"classes/_store_upgrade_.upgrade.html#newversion","classes":"tsd-kind-property tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":91,"kind":512,"name":"constructor","url":"classes/_store_upgrade_.upgrade.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"store/Upgrade\".Upgrade"},{"id":92,"kind":2048,"name":"createStore","url":"classes/_store_upgrade_.upgrade.html#createstore","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Upgrade\".Upgrade"},{"id":93,"kind":2048,"name":"transaction","url":"classes/_store_upgrade_.upgrade.html#transaction","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Upgrade\".Upgrade"},{"id":94,"kind":1,"name":"\"store/OpenOptions\"","url":"modules/_store_openoptions_.html","classes":"tsd-kind-external-module"},{"id":95,"kind":256,"name":"OpenOptions","url":"interfaces/_store_openoptions_.openoptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/OpenOptions\""},{"id":96,"kind":2048,"name":"upgrade","url":"interfaces/_store_openoptions_.openoptions.html#upgrade","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"\"store/OpenOptions\".OpenOptions"},{"id":97,"kind":2048,"name":"blocked","url":"interfaces/_store_openoptions_.openoptions.html#blocked","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"\"store/OpenOptions\".OpenOptions"},{"id":98,"kind":2048,"name":"blocking","url":"interfaces/_store_openoptions_.openoptions.html#blocking","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"\"store/OpenOptions\".OpenOptions"},{"id":99,"kind":1024,"name":"autoCloseOnBlocking","url":"interfaces/_store_openoptions_.openoptions.html#autocloseonblocking","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"store/OpenOptions\".OpenOptions"},{"id":100,"kind":1,"name":"\"store/Database\"","url":"modules/_store_database_.html","classes":"tsd-kind-external-module"},{"id":101,"kind":4,"name":"TransactionMode","url":"enums/_store_database_.transactionmode.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"store/Database\""},{"id":102,"kind":16,"name":"ReadOnly","url":"enums/_store_database_.transactionmode.html#readonly","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"store/Database\".TransactionMode"},{"id":103,"kind":16,"name":"ReadWrite","url":"enums/_store_database_.transactionmode.html#readwrite","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"store/Database\".TransactionMode"},{"id":104,"kind":128,"name":"Database","url":"classes/_store_database_.database.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"store/Database\""},{"id":105,"kind":2048,"name":"open","url":"classes/_store_database_.database.html#open","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter tsd-is-static","parent":"\"store/Database\".Database"},{"id":106,"kind":2048,"name":"put","url":"classes/_store_database_.database.html#put","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":107,"kind":2048,"name":"get","url":"classes/_store_database_.database.html#get","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":108,"kind":2048,"name":"transaction","url":"classes/_store_database_.database.html#transaction","classes":"tsd-kind-method tsd-parent-kind-class tsd-has-type-parameter","parent":"\"store/Database\".Database"},{"id":109,"kind":2048,"name":"close","url":"classes/_store_database_.database.html#close","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"store/Database\".Database"}]};
