+++ docs/api/enums/_store_database_.transactionmode.html
@@ -91,7 +91,7 @@
 					<div class="tsd-signature tsd-kind-icon">Read<wbr>Only<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;readonly&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/LBHackney-IT/remultiform/blob/master/src/store/Database.ts#L18">store/Database.ts:18</a></li>
+							<li>Defined in <a href="https://github.com/LBHackney-IT/remultiform/blob/master/src/store/Database.ts#L17">store/Database.ts:17</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -101,7 +101,7 @@
 					<div class="tsd-signature tsd-kind-icon">Read<wbr>Write<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;readwrite&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/LBHackney-IT/remultiform/blob/master/src/store/Database.ts#L19">store/Database.ts:19</a></li>
+							<li>Defined in <a href="https://github.com/LBHackney-IT/remultiform/blob/master/src/store/Database.ts#L18">store/Database.ts:18</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -124,6 +124,9 @@
 					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="../modules/_store_database_.html">&quot;store/<wbr>Database&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
