+++ docs/api/globals.html
@@ -74,6 +74,7 @@
 								<li class="tsd-kind-external-module"><a href="modules/_components_page_.html" class="tsd-kind-icon">&quot;components/<wbr>Page&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_helpers_pagecomponentwrapper_.html" class="tsd-kind-icon">&quot;helpers/<wbr>Page<wbr>Component<wbr>Wrapper&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_store_database_.html" class="tsd-kind-icon">&quot;store/<wbr>Database&quot;</a></li>
+								<li class="tsd-kind-external-module"><a href="modules/_store_openoptions_.html" class="tsd-kind-icon">&quot;store/<wbr>Open<wbr>Options&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_store_upgrade_.html" class="tsd-kind-icon">&quot;store/<wbr>Upgrade&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_store_types_.html" class="tsd-kind-icon">&quot;store/types&quot;</a></li>
 							</ul>
@@ -99,6 +100,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_store_database_.html">&quot;store/<wbr>Database&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="modules/_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
