+++ docs/api/interfaces/_components_orchestrator_.step.html
@@ -134,6 +134,9 @@
 						<a href="../modules/_store_database_.html">&quot;store/<wbr>Database&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
+						<a href="../modules/_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
 						<a href="../modules/_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
