+++ docs/api/modules/_store_openoptions_.html
@@ -3,7 +3,7 @@
 <head>
 	<meta charset="utf-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<title>Schema | LBH Frontend React</title>
+	<title>&quot;store/OpenOptions&quot; | LBH Frontend React</title>
 	<meta name="description" content="">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="stylesheet" href="../assets/css/main.css">
@@ -56,46 +56,28 @@
 					<a href="../globals.html">Globals</a>
 				</li>
 				<li>
-					<a href="../modules/_store_types_.html">&quot;store/types&quot;</a>
-				</li>
-				<li>
-					<a href="_store_types_.schema.html">Schema</a>
+					<a href="_store_openoptions_.html">&quot;store/OpenOptions&quot;</a>
 				</li>
 			</ul>
-			<h1>Interface Schema</h1>
+			<h1>External module &quot;store/OpenOptions&quot;</h1>
 		</div>
 	</div>
 </header>
 <div class="container container-main">
 	<div class="row">
 		<div class="col-8 col-content">
-			<section class="tsd-panel tsd-comment">
-				<div class="tsd-comment tsd-typography">
-					<div class="lead">
-						<p>A database schema.</p>
+			<section class="tsd-panel-group tsd-index-group">
+				<h2>Index</h2>
+				<section class="tsd-panel tsd-index-panel">
+					<div class="tsd-index-content">
+						<section class="tsd-index-section ">
+							<h3>Interfaces</h3>
+							<ul class="tsd-index-list">
+								<li class="tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter"><a href="../interfaces/_store_openoptions_.openoptions.html" class="tsd-kind-icon">Open<wbr>Options</a></li>
+							</ul>
+						</section>
 					</div>
-					<p>IndexedDB databases are made up of stores, mapping store name to
-					<a href="_store_types_.storeschema.html">StoreSchema</a>.</p>
-				</div>
-			</section>
-			<section class="tsd-panel tsd-hierarchy">
-				<h3>Hierarchy</h3>
-				<ul class="tsd-hierarchy">
-					<li>
-						<span class="target">Schema</span>
-					</li>
-				</ul>
-			</section>
-			<section class="tsd-panel tsd-kind-interface tsd-parent-kind-external-module">
-				<h3 class="tsd-before-signature">Indexable</h3>
-				<div class="tsd-signature tsd-kind-icon"><span class="tsd-signature-symbol">[</span>storeName: <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">]: </span><a href="_store_types_.storeschema.html" class="tsd-signature-type">StoreSchema</a></div>
-				<div class="tsd-comment tsd-typography">
-					<div class="lead">
-						<p>A database schema.</p>
-					</div>
-					<p>IndexedDB databases are made up of stores, mapping store name to
-					<a href="_store_types_.storeschema.html">StoreSchema</a>.</p>
-				</div>
+				</section>
 			</section>
 		</div>
 		<div class="col-4 col-menu menu-sticky-wrap menu-highlight">
@@ -105,60 +87,32 @@
 						<a href="../globals.html"><em>Globals</em></a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="../modules/_components_orchestrator_.html">&quot;components/<wbr>Orchestrator&quot;</a>
+						<a href="_components_orchestrator_.html">&quot;components/<wbr>Orchestrator&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="../modules/_components_page_.html">&quot;components/<wbr>Page&quot;</a>
+						<a href="_components_page_.html">&quot;components/<wbr>Page&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="../modules/_helpers_pagecomponentwrapper_.html">&quot;helpers/<wbr>Page<wbr>Component<wbr>Wrapper&quot;</a>
+						<a href="_helpers_pagecomponentwrapper_.html">&quot;helpers/<wbr>Page<wbr>Component<wbr>Wrapper&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="../modules/_store_database_.html">&quot;store/<wbr>Database&quot;</a>
-					</li>
-					<li class=" tsd-kind-external-module">
-						<a href="../modules/_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
-					</li>
-					<li class=" tsd-kind-external-module">
-						<a href="../modules/_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
+						<a href="_store_database_.html">&quot;store/<wbr>Database&quot;</a>
 					</li>
 					<li class="current tsd-kind-external-module">
-						<a href="../modules/_store_types_.html">&quot;store/types&quot;</a>
+						<a href="_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_store_types_.html">&quot;store/types&quot;</a>
 					</li>
 				</ul>
 			</nav>
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="_store_types_.namedschema.html" class="tsd-kind-icon">Named<wbr>Schema</a>
-					</li>
-				</ul>
-				<ul class="current">
-					<li class="current tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_store_types_.schema.html" class="tsd-kind-icon">Schema</a>
-					</li>
-				</ul>
-				<ul class="after-current">
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_store_types_.storeschema.html" class="tsd-kind-icon">Store<wbr>Schema</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#store" class="tsd-kind-icon">Store</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#storekey" class="tsd-kind-icon">Store<wbr>Key</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#storemap" class="tsd-kind-icon">Store<wbr>Map</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#storenames" class="tsd-kind-icon">Store<wbr>Names</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#storevalue" class="tsd-kind-icon">Store<wbr>Value</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../modules/_store_types_.html#transaction" class="tsd-kind-icon">Transaction</a>
+						<a href="../interfaces/_store_openoptions_.openoptions.html" class="tsd-kind-icon">Open<wbr>Options</a>
 					</li>
 				</ul>
 			</nav>
