+++ docs/api/modules/_store_upgrade_.html
@@ -98,6 +98,9 @@
 					<li class=" tsd-kind-external-module">
 						<a href="_store_database_.html">&quot;store/<wbr>Database&quot;</a>
 					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_store_openoptions_.html">&quot;store/<wbr>Open<wbr>Options&quot;</a>
+					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="_store_upgrade_.html">&quot;store/<wbr>Upgrade&quot;</a>
 					</li>
